pyproject.toml
# Modern Python project configuration for Automotive Claude Code
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automotive-claude-code"
version = "0.1.0"
description = "Universal AI-powered development platform for automotive industry"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Automotive Claude Code Contributors", email = "automotive-claude-code@opensource.ai"}
]
maintainers = [
    {name = "Automotive Claude Code Contributors", email = "automotive-claude-code@opensource.ai"}
]
keywords = [
    "automotive",
    "autosar",
    "adas",
    "battery",
    "ev",
    "autonomous",
    "safety",
    "iso26262",
    "ai",
    "llm",
    "claude",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Automotive",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.54.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.0",
    "python-can>=4.4.2",
    "cantools>=39.4.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.8",
    "black>=24.8.0",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
    "pre-commit>=3.5.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.39",
    "mkdocstrings>=0.24.0",
    "mkdocs-mermaid2-plugin>=1.1.1",
]
adas = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "open3d>=0.17.0",
]
battery = [
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
]
diagnostic = [
    "udsoncan>=1.21.0",
    "doip>=1.2.0",
]
all = [
    "automotive-claude-code[dev,docs,adas,battery,diagnostic]",
]

[project.urls]
Homepage = "https://github.com/automotive-opensource/automotive-claude-code-agents"
Documentation = "https://automotive-claude-code.readthedocs.io/"
Repository = "https://github.com/automotive-opensource/automotive-claude-code-agents"
Issues = "https://github.com/automotive-opensource/automotive-claude-code-agents/issues"
Changelog = "https://github.com/automotive-opensource/automotive-claude-code-agents/blob/main/CHANGELOG.md"

[project.scripts]
automotive-cursor-setup = "tools.cursor_setup:main"
automotive-llm-council = "tools.llm_council:run_sync_cli"

[tool.setuptools]
packages = ["tools", "tools.adapters"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.md"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["ARG", "S101"]

[tool.ruff.lint.isort]
known-first-party = ["tools"]

# MyPy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--cov=tools",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
    "can: CAN bus related tests",
    "autosar: AUTOSAR related tests",
    "battery: Battery management tests",
    "adas: ADAS/perception tests",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["tools"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

# Bandit configuration (security)
[tool.bandit]
exclude_dirs = ["tests", "venv", "env"]
skips = ["B101", "B601"]

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
